requirements.txt
channels==3.0.4
dialogflow==0.5.1
dialogflow-fulfillment==0.4.5
Django==4.2
django-allauth==0.52.0
django-environ==0.10.0
google-api-core==1.4.1
google-auth==1.35.0
googleapis-common-protos==1.59.0
requests==2.28.2
stripe==2.63.0
validate-email==1.3
Pillow==9.5.0